% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cau.r
\name{Cau}
\alias{Cau}
\title{compute the cdf of the Cauchy distribution}
\usage{
Cau(theta, eta, y)
}
\arguments{
\item{theta}{scale parameter}

\item{eta}{location parameter}

\item{y}{vector of quantiles.}
}
\value{
a random sample of size \code{n}
}
\description{
Use a function to compute the cdf of the Cauchy distribution, theta and eta are scale and location parameter respectively
}
\examples{
\dontrun{
theta = 3
eta = 2
y <- seq(-10, 10, length(10))
Cau(theta, eta, y)
}
}
