% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.Phi.r
\name{MC.Phi}
\alias{MC.Phi}
\title{generate samples from a Rayleigh distribution.}
\usage{
MC.Phi(x, sigma, r, antithetic = TRUE)
}
\arguments{
\item{x}{vector of quantiles}

\item{sigma}{a parameter}

\item{r}{number of sample}

\item{antithetic}{logical value, the default is true}
}
\value{
a vector of size x\code{n}
}
\description{
Implement a function to generate samples from a Rayleigh distribution.
}
\examples{
\dontrun{
x <- seq(1, 10, 1)
sigma <- 2
r <- 10000
res <- MC.Phi(x, sigma, r)
}
}
