% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCe.r
\name{MCe}
\alias{MCe}
\title{compute Monte Carlo estimate of the Beta(3,3) cdf.}
\usage{
MCe(x, n)
}
\arguments{
\item{x}{vector of quantiles}

\item{n}{number of random sampling}
}
\value{
a vector of size x\code{n}
}
\description{
Use a function to compute Monte Carlo estimate of the Beta(3,3) cdf.
}
\examples{
\dontrun{
x <- seq(0.1, 0.9, 0.1)
n <- 10000
res <- MCe(x, n)
}
}
